Вынес списко студентов в отдельный компонент
